+++ CHANGES.md
@@ -10,4 +10,6 @@
 
 ## Behavioral changes and enhancements
 
+* [SSHD-930](https://issues.apache.org/jira/browse/SSHD-930) - Added configuration allowing the user to specify whether client should wait
+for the server's identification before sending its own.